refactor(user): 서비스 패키지 구조 정리 (#54)
- 서비스 클래스들을 application/service/ 디렉토리로 이동
  - JwtTokenService: jwt/service/ → application/service/
  - EncryptionService: encryption/service/ → application/service/
  - EmailService, EmailSenderService: email/service/ → application/service/
  - StudentNumberGenerator → StudentNumberGeneratorService: domain/service/ → application/service/

- 중복 인터페이스 EmailVerificationService 제거
  - EmailVerificationServiceImpl만 유지

- 관련 import 문 업데이트
  - AttendanceService, DataEncryptionAdapter
  - JwtAuthenticationFilter, EncryptedFieldSerializer
  - SignupService, UserViewServiceImpl
  - ProfessorViewServiceImpl, StudentViewServiceImpl

- ArchUnit freeze store 업데이트
