--- springboot/archunit_store/76c6a14c-b3b3-4293-886c-5e862636f164
@@ -1,11 +1,0 @@
-Constructor <com.mzc.backend.lms.domains.user.domain.service.StudentNumberGenerator.<init>(com.mzc.backend.lms.domains.user.adapter.out.persistence.repository.StudentNumberSequenceRepository, com.mzc.backend.lms.domains.user.adapter.out.persistence.repository.UserRepository)> has parameter of type <com.mzc.backend.lms.domains.user.adapter.out.persistence.repository.StudentNumberSequenceRepository> in (StudentNumberGenerator.java:0)
-Constructor <com.mzc.backend.lms.domains.user.domain.service.StudentNumberGenerator.<init>(com.mzc.backend.lms.domains.user.adapter.out.persistence.repository.StudentNumberSequenceRepository, com.mzc.backend.lms.domains.user.adapter.out.persistence.repository.UserRepository)> has parameter of type <com.mzc.backend.lms.domains.user.adapter.out.persistence.repository.UserRepository> in (StudentNumberGenerator.java:0)
-Field <com.mzc.backend.lms.domains.user.domain.service.StudentNumberGenerator.sequenceRepository> has type <com.mzc.backend.lms.domains.user.adapter.out.persistence.repository.StudentNumberSequenceRepository> in (StudentNumberGenerator.java:0)
-Field <com.mzc.backend.lms.domains.user.domain.service.StudentNumberGenerator.userRepository> has type <com.mzc.backend.lms.domains.user.adapter.out.persistence.repository.UserRepository> in (StudentNumberGenerator.java:0)
-Method <com.mzc.backend.lms.domains.user.domain.service.StudentNumberGenerator.createSequenceWithExistingData(java.lang.Integer, java.lang.Long, java.lang.Long)> calls method <com.mzc.backend.lms.domains.user.adapter.out.persistence.entity.StudentNumberSequence.createWithInitialSequence(java.lang.Integer, java.lang.Long, java.lang.Long, java.lang.Integer)> in (StudentNumberGenerator.java:97)
-Method <com.mzc.backend.lms.domains.user.domain.service.StudentNumberGenerator.createSequenceWithExistingData(java.lang.Integer, java.lang.Long, java.lang.Long)> calls method <com.mzc.backend.lms.domains.user.adapter.out.persistence.repository.StudentNumberSequenceRepository.save(java.lang.Object)> in (StudentNumberGenerator.java:99)
-Method <com.mzc.backend.lms.domains.user.domain.service.StudentNumberGenerator.createSequenceWithExistingData(java.lang.Integer, java.lang.Long, java.lang.Long)> calls method <com.mzc.backend.lms.domains.user.adapter.out.persistence.repository.UserRepository.findMaxSequenceByPrefix(java.lang.String)> in (StudentNumberGenerator.java:91)
-Method <com.mzc.backend.lms.domains.user.domain.service.StudentNumberGenerator.createSequenceWithExistingData(java.lang.Integer, java.lang.Long, java.lang.Long)> has return type <com.mzc.backend.lms.domains.user.adapter.out.persistence.entity.StudentNumberSequence> in (StudentNumberGenerator.java:0)
-Method <com.mzc.backend.lms.domains.user.domain.service.StudentNumberGenerator.generateNumber(java.lang.Long, java.lang.Long)> calls method <com.mzc.backend.lms.domains.user.adapter.out.persistence.entity.StudentNumberSequence.getNextSequence()> in (StudentNumberGenerator.java:68)
-Method <com.mzc.backend.lms.domains.user.domain.service.StudentNumberGenerator.generateNumber(java.lang.Long, java.lang.Long)> calls method <com.mzc.backend.lms.domains.user.adapter.out.persistence.repository.StudentNumberSequenceRepository.findByYearAndCollegeAndDepartmentWithLock(java.lang.Integer, java.lang.Long, java.lang.Long)> in (StudentNumberGenerator.java:64)
-Method <com.mzc.backend.lms.domains.user.domain.service.StudentNumberGenerator.generateNumber(java.lang.Long, java.lang.Long)> calls method <com.mzc.backend.lms.domains.user.adapter.out.persistence.repository.StudentNumberSequenceRepository.save(java.lang.Object)> in (StudentNumberGenerator.java:69)
