+++ springboot/archunit_store/92afd9a8-6d40-404a-b946-a65aec825a0d
@@ -2,4 +2,3 @@ Class <com.mzc.backend.lms.domains.enrollment.application.service.CartUseCaseImp
 Class <com.mzc.backend.lms.domains.enrollment.application.service.EnrollmentCourseUseCaseImpl> does not have simple name ending with 'Service' in (EnrollmentCourseUseCaseImpl.java:0) and Class <com.mzc.backend.lms.domains.enrollment.application.service.EnrollmentCourseUseCaseImpl> does not have simple name ending with 'ServiceImpl' in (EnrollmentCourseUseCaseImpl.java:0)
 Class <com.mzc.backend.lms.domains.enrollment.application.service.EnrollmentPeriodUseCaseImpl> does not have simple name ending with 'Service' in (EnrollmentPeriodUseCaseImpl.java:0) and Class <com.mzc.backend.lms.domains.enrollment.application.service.EnrollmentPeriodUseCaseImpl> does not have simple name ending with 'ServiceImpl' in (EnrollmentPeriodUseCaseImpl.java:0)
 Class <com.mzc.backend.lms.domains.enrollment.application.service.EnrollmentUseCaseImpl> does not have simple name ending with 'Service' in (EnrollmentUseCaseImpl.java:0) and Class <com.mzc.backend.lms.domains.enrollment.application.service.EnrollmentUseCaseImpl> does not have simple name ending with 'ServiceImpl' in (EnrollmentUseCaseImpl.java:0)
-Class <com.mzc.backend.lms.domains.user.domain.service.StudentNumberGenerator> does not have simple name ending with 'Service' in (StudentNumberGenerator.java:0) and Class <com.mzc.backend.lms.domains.user.domain.service.StudentNumberGenerator> does not have simple name ending with 'ServiceImpl' in (StudentNumberGenerator.java:0)
