+++ springboot/archunit_store/9f8734e2-5f2c-43bd-bd48-ca7758123696
@@ -1,0 +1,1 @@
+Class <com.mzc.backend.lms.domains.notification.adapter.out.queue.service.NotificationQueueServiceImpl> does not reside in a package '..application.service..' in (NotificationQueueServiceImpl.java:0)
