+++ springboot/archunit_store/e14d00a8-110b-46b3-bd0d-7dee253d0d2f
