+++ springboot/src/main/java/com/mzc/backend/lms/domains/user/application/service/EmailService.java
@@ -1,4 +1,4 @@
-package com.mzc.backend.lms.domains.user.email.service;
+package com.mzc.backend.lms.domains.user.application.service;
 
 import com.mzc.backend.lms.domains.user.email.dto.EmailMessage;
 import com.mzc.backend.lms.domains.user.exception.AuthException;
